RexStan und php-cs-fixer

+++ lib/GeoCoder.php
@@ -262,7 +262,7 @@ class GeoCoder
          */
         $value = trim(rex_request::request(self::VALUE_PARAM, 'string', ''));
         if ('' === $value) {
-            $ttl = round((strtotime('today midnight') - time()) / 60, 0);
+            $ttl = (int) round((strtotime('today midnight') - time()) / 60, 0);
             Tools::sendJson([], time(), $ttl);
         }
 
@@ -272,7 +272,7 @@ class GeoCoder
          */
         $geoCoder = self::get($geoCoderId);
         if (null === $geoCoder) {
-            $ttl = round((strtotime('today midnight') - time()) / 60, 0);
+            $ttl = (int) round((strtotime('today midnight') - time()) / 60, 0);
             Tools::sendJson([], time(), $ttl);
         }
 
@@ -315,7 +315,7 @@ class GeoCoder
          * Beim Versand wird als TimeToLive die Zeit von Jetzt bis Mitternacht
          * eingestellt (Angabe in Minuten).
          */
-        $ttl = round((strtotime('today midnight') - time()) / 60, 0);
+        $ttl = (int) round((strtotime('today midnight') - time()) / 60, 0);
         Tools::sendJson(json_decode($content, true) ?? [], time(), $ttl);
     }
